backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg2-binary
pydantic
pillow
anyio
openai
python-multipart